+++ unique_toolkit/unique_toolkit/services/knowledge_base.py
@@ -303,6 +303,7 @@ class KnowledgeBaseService:
         self,
         *,
         where: dict,
+        include_failed_content: bool = False,
     ) -> list[Content]:
         """
         Performs a search in the knowledge base by filter (and not a smilarity search)
@@ -320,12 +321,14 @@ class KnowledgeBaseService:
             company_id=self._company_id,
             chat_id="",
             where=where,
+            include_failed_content=include_failed_content,
         )
 
     async def search_contents_async(
         self,
         *,
         where: dict,
+        include_failed_content: bool = False,
     ) -> list[Content]:
         """
         Performs an asynchronous search for content files in the knowledge base by filter.
@@ -342,6 +345,7 @@ class KnowledgeBaseService:
             company_id=self._company_id,
             chat_id="",
             where=where,
+            include_failed_content=include_failed_content,
         )
 
     def upload_content_from_bytes(